+++ apps/client/src/components/common/ProductCard/index.ts
@@ -1,0 +1,9 @@
+import Main from './Main';
+import Order from './Order';
+
+const ProductCard = {
+    Main,
+    Order,
+};
+
+export default ProductCard;
