+++ apps/client/src/components/features/my-order/OrderedProductCard.tsx
@@ -1,4 +1,4 @@
-import ProductCard from '@/components/common/ProductCard/Order';
+import ProductCard from '@/components/common/ProductCard';
 import { OrderedObjectDetailInfo } from '@/schemas/order';
 
 const OrderedProductCard = ({
@@ -7,17 +7,17 @@ const OrderedProductCard = ({
     object: OrderedObjectDetailInfo;
 }) => {
     return (
-        <ProductCard.Root>
-            <ProductCard.Image
+        <ProductCard.Order.Root>
+            <ProductCard.Order.Image
                 imageUrl={object.imageUrl}
                 alt={object.objectName}
             />
-            <ProductCard.Info
+            <ProductCard.Order.Info
                 name={object.objectName}
                 intro={object.intro}
                 price={object.purchasePrice}
             />
-        </ProductCard.Root>
+        </ProductCard.Order.Root>
     );
 };
